requirements.txt
fastapi==0.109.0
uvicorn==0.23.2
alembic==1.13.1
pydantic==2.5.3
SQLAlchemy==2.0.25
sqlmodel==0.0.14
dependency-injector==4.41.0
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4
psycopg2==2.9.9
pytest==7.4.2
httpx==0.25.0
flake8==6.1.0
pre-commit==3.5.0
flake8-docstrings==1.7.0
flake8-import-order==0.18.2